requirements.txt
# EdgeTX Tuning Chirp – install with: pip install -r requirements.txt
numpy
matplotlib
scipy
PyYAML
